Rename _selection_strategy.py as _variants.py.
