+++ src/kaczmarz/_variants.py
